requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.35
aiogram==3.22.0
python-dotenv==1.0.1
pytz==2024.2
yookassa==3.3.0
aiohttp==3.10.10
werkzeug==3.0.6
PyJWT==2.8.0
python-multipart==0.0.12
schedule==1.2.2
aiofiles==24.1.0
psutil==5.9.8
redis==5.0.1
Pillow==10.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2